Refactor out testNonConcurrentSnapshotTransformerAccess to separte StateSnapshotTransformerTest
